# src/main/resources/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: DoeAqui API
  description: API atômica para o sistema de doação de sangue DoeAqui.
  version: 1.0.0
servers:
  - url: http://localhost:8080/v1
paths:
...
  /hemocentro:
    post:
      summary: Cadastrar novo hemocentro
      operationId: postNewHemocentro
      tags:
        - Hemocentro
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewHemocentroRequest'
      responses:
        '201':
          description: Hemocentro cadastrado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HemocentroResponse'
        '400':
          description: Requisição inválida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
        '500':
          description: Erro no servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'

    get:
      summary: Obter informações de hemocentro por filtros
      operationId: getHemocentroByFilter
      tags:
        - Hemocentro
      parameters:
        - in: query
          name: nome
          schema:
            type: string
            maxLength: 150
          required: false
        - in: query
          name: telefone
          schema:
            type: string
            maxLength: 20
          required: false
        - in: query
          name: email
          schema:
            type: string
            format: email
            maxLength: 100
          required: false
      responses:
        '200':
          description: Hemocentro encontrado com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HemocentroResponse'
        '400':
          description: Requisição inválida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
        '500':
          description: Erro no servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
  /hemocentro/{idHemocentro}:
    get:
      summary: Obter informações de um hemocentro por ID
      operationId: getHemocentroInfoById
      tags:
        - Hemocentro
      parameters:
        - name: idHemocentro
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Informações do hemocentro obtidas com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HemocentroResponse'
        '400':
          description: Requisição inválida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
        '404':
          description: Hemocentro não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
        '500':
          description: Erro no servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'

    patch:
      summary: Atualizar dados de um hemocentro
      operationId: patchHemocentroInfo
      tags:
        - Hemocentro
      parameters:
        - name: idHemocentro
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateHemocentroRequest'
      responses:
        '200':
          description: Hemocentro atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HemocentroResponse'
        '400':
          description: Requisição inválida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
        '404':
          description: Hemocentro não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
        '500':
          description: Erro no servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'

    delete:
      summary: Inativar o cadastro de um hemocentro
      operationId: deleteHemocentro
      tags:
        - Hemocentro
      parameters:
        - name: idHemocentro
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Hemocentro inativado com sucesso
        '400':
          description: Requisição inválida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
        '404':
          description: Hemocentro não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
        '500':
          description: Erro no servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
  /hemocentro/proximo:
    get:
      summary: Obter hemocentros mais próximos do usuário
      operationId: getHemocentroByLocation
      tags:
        - Hemocentro
      parameters:
        - in: query
          name: latitude
          schema:
            type: number
            format: double
          required: true
        - in: query
          name: longitude
          schema:
            type: number
            format: double
          required: true
        - in: query
          name: raio
          schema:
            type: integer
            format: int32
          required: true
      responses:
        '200':
          description: Hemocentros encontrados com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HemocentroPorLocalizacaoResponse'
        '400':
          description: Requisição inválida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
        '500':
          description: Erro no servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
...
components:
  schemas:
...
    NewHemocentroRequest:
      type: object
      required:
        - nome
        - telefone
        - email
        - endereco
      properties:
        nome:
          type: string
          minLength: 1
          maxLength: 150
          pattern: "^[a-zA-ZÀ-ú'\\s]+$"
        telefone:
          type: string
          minLength: 1
          maxLength: 20
          pattern: '^[0-9]{10,11}$'
        email:
          type: string
          format: email
          minLength: 1
          maxLength: 100
          pattern: '^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$'
        endereco:
          $ref: '#/components/schemas/NewEnderecoRequest'
    HemocentroResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
        nome:
          type: string
          minLength: 1
          maxLength: 150
        telefone:
          type: string
          maxLength: 20
        email:
          type: string
          format: email
          maxLength: 100
        ativo:
          type: boolean
        endereco:
          $ref: '#/components/schemas/Endereco'
    HemocentroPorLocalizacaoResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
        nome:
          type: string
          minLength: 1
          maxLength: 150
        telefone:
          type: string
          maxLength: 20
        email:
          type: string
          format: email
          maxLength: 100
        prioridade:
          type: boolean
        endereco:
          $ref: '#/components/schemas/Endereco'
...
    UpdateHemocentroRequest:
      type: object
      properties:
        telefone:
          type: string
          minLength: 1
          maxLength: 20
          pattern: '^[0-9]{10,11}$'
        email:
          type: string
          format: email
          minLength: 1
          maxLength: 100
          pattern: '^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$'
        ativo:
          type: boolean
...
    NewEnderecoRequest:
      type: object
      required:
        - rua
        - bairro
        - numero
        - cep
        - cidade
        - estado
        - pais
        - coordenadas
      properties:
        rua:
          type: string
          minLength: 1
          maxLength: 255
        bairro:
          type: string
          minLength: 1
          maxLength: 100
        numero:
          type: string
          minLength: 1
          maxLength: 10
        complemento:
          type: string
          minLength: 1
          maxLength: 20
        cep:
          type: string
          pattern: '^[0-9]{8}$'
          minLength: 8
          maxLength: 8
        cidade:
          type: string
          minLength: 1
          maxLength: 100
        estado:
          type: string
          pattern: '^[a-zA-Z]{2}$'
          minLength: 2
          maxLength: 2
        pais:
          type: string
          minLength: 1
          maxLength: 100
        coordenadas:
          type: object
          required:
            - longitude
            - latitude
          properties:
            longitude:
              type: number
              format: double
            latitude:
              type: number
              format: double
    Endereco:
      type: object
      properties:
        rua:
          type: string
        bairro:
          type: string
        numero:
          type: string
        complemento:
          type: string
        cep:
          type: string
        cidade:
          type: string
        estado:
          type: string
        pais:
          type: string
        coordenadas:
          type: object
          properties:
            longitude:
              type: number
              format: double
            latitude:
              type: number
              format: double
    StandardError:
      type: object
      required:
        - timestamp
        - status
        - error
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string